backend/requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.2
gunicorn==23.0.0
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
uvicorn==0.35.0
Werkzeug==3.1.3
pymongo==4.6.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
flask-cors==4.0.0
requests==2.31.0

opencv-python==4.8.1.78
mediapipe>=0.10.0
numpy==1.24.3
pillow==10.0.1
pytesseract==0.3.10
easyocr==1.7.0
scikit-learn==1.3.2
scipy==1.11.4